Remove impls file and move extend to mod
